requirements.txt
pyfiglet
PyNaCl